Refactored activity names and moved them to their own package
