+++ app/src/main/AndroidManifest.xml
@@ -14,21 +14,21 @@
         android:theme="@style/Theme.ResDesUx2"
         tools:targetApi="31">
         <activity
-            android:name=".calendar"
+            android:name=".Activities.CalendarActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".Dashboard"
+            android:name=".Activities.DashboardActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".Activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
